+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 - [247](https://github.com/vegaprotocol/vegatools/issues/247) - Made ganache value optional for perftool
 - [251](https://github.com/vegaprotocol/vegatools/issues/251) - Changed code to use wallet V2 and updated protobuf mod to correct version
 - [254](https://github.com/vegaprotocol/vegatools/issues/254) - Updated auth to use VWT header value
+- [256](https://github.com/vegaprotocol/vegatools/issues/256) - Fix batch orders 
 
 ### 🐛 Fixes
 - [78](https://github.com/vegaprotocol/vegatools/pull/78) - Fix build with missing dependency